config/cdk/requirements.txt
aws-cdk.aws-ec2
aws-cdk.aws-rds
aws-cdk.aws_elasticache
aws-cdk.aws-s3
aws-cdk.aws-autoscaling
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-autoscaling-common
aws-cdk.aws-iam